moved around tests, fixed log output
