+++ test/src/hello.go
@@ -35,5 +35,4 @@ func main() {
 	res, _ := json.Marshal(msg)
 	fmt.Printf("name=%s\n", name)
 	fmt.Println(string(res))
-
 }
